+++ src/CFG.js
@@ -5,11 +5,6 @@ const CFG = {};
     // Boolean for verbose reporting
     'DEBUG', // Effectively defaults to false (ignored unless `true`)
 
-    // Used when setting global shims to determine whether to try to add
-    //   other globals shimmed by the library (`ShimDOMException`, `ShimDOMStringList`,
-    //   `ShimEvent`, `ShimCustomEvent`, `ShimEventTarget`)
-    'addNonIDBGlobals', // Effectively defaults to false (ignored unless `true`)
-
     // Determines whether the slow-performing `Object.setPrototypeOf` calls required
     //    for full WebIDL compliance will be used. Probably only needed for testing
     //    or environments where full introspection on class relationships is required;
@@ -28,14 +23,6 @@ const CFG = {};
     //    expression strings of `src/UnicodeIdentifiers.js`)
     'UnicodeIDStart', // In the non-Unicode builds, defaults to /[$A-Z_a-z]/
     'UnicodeIDContinue', // In the non-Unicode builds, defaults to /[$0-9A-Z_a-z]/
-
-    // NODE-SPECIFIC CONFIG
-    // Boolean on whether to delete the database file itself after `deleteDatabase`;
-    //   defaults to `true` as the database will be empty
-    'deleteDatabaseFiles',
-    // Boolean on whether to add the `.sqlite` extension to file names;
-    //   defaults to `true`
-    'addSQLiteExtension',
 
     // -----------SQL CONFIG----------
     // Object (`window` in the browser) on which there may be an
@@ -56,6 +43,14 @@ const CFG = {};
     //  quota every five megabytes."
     'DEFAULT_DB_SIZE', // Defaults to (4 * 1024 * 1024) or (25 * 1024 * 1024) in Safari
 
+    // NODE-IMPINGING SETTINGS (created for sake of limitations in Node or desktop file
+    //    system implementation but applied by default in browser for parity)
+
+    // Used when setting global shims to determine whether to try to add
+    //   other globals shimmed by the library (`ShimDOMException`, `ShimDOMStringList`,
+    //   `ShimEvent`, `ShimCustomEvent`, `ShimEventTarget`)
+    'addNonIDBGlobals', // Effectively defaults to false (ignored unless `true`)
+
     // Overcoming limitations with node-sqlite3/storing database name on file systems
     // https://en.wikipedia.org/wiki/Filename#Reserved_characters_and_words
     'escapeDatabaseName', // Defaults to prefixing database with `D_`, escaping
@@ -67,8 +62,18 @@ const CFG = {};
         // (characters nevertheless commonly reserved in modern, Unicode-supporting
         // systems): 0x00-0x1F 0x7F " * / : < > ? \ |
     'databaseNameLengthLimit', // Defaults to 254 (shortest typical modern file length limit)
+    'escapeNFDForDatabaseNames', // Boolean defaulting to true on whether to escape NFD-escaping
+                                // characters to avoid clashes on MacOS which performs NFD on files
+    // Boolean on whether to add the `.sqlite` extension to file names;
+    //   defaults to `true`
+    'addSQLiteExtension',
 
-    // Optional Node WebSQL config
+    // NODE-SPECIFIC CONFIG
+    // Boolean on whether to delete the database file itself after `deleteDatabase`;
+    //   defaults to `true` as the database will be empty
+    'deleteDatabaseFiles',
+
+    // NODE-SPECIFIC WEBSQL CONFIG
     'sqlBusyTimeout', // Defaults to 1000
     'sqlTrace', // Callback not used by default
     'sqlProfile' // Callback not used by default