+++ tests-mocha/test-utils.js
@@ -5,7 +5,7 @@
     'use strict';
 
     var databaseNamePrefix = 'indexeddbshim_test_database_';
-    var databaseSuffix = '.sqlite';
+    var databaseSuffix = ''; // '.sqlite' now added automatically by default
     var dbNameCounter = 0;
     var util;