+++ tests-polyfill/w3c/IDBFactory.deleteDatabase.js
@@ -64,7 +64,7 @@ describe('W3C IDBFactory.deleteDatabase Tests', function () {
     // idbfactory_deletedatabase4
     it('Test events opening a second database when one connection is open already', function (done) {
         var db;
-        var openrq = indexedDB.open('db.sqlite', 3);
+        var openrq = indexedDB.open('db', 3);
 
         openrq.onupgradeneeded = function(e) {
             e.target.result.createObjectStore('store');
@@ -86,7 +86,7 @@ describe('W3C IDBFactory.deleteDatabase Tests', function () {
         openrq.onblocked = function () { throw new Error("open.blocked"); };
 
         function Second(e) {
-            var deleterq = indexedDB.deleteDatabase('db.sqlite');
+            var deleterq = indexedDB.deleteDatabase('db');
 
             deleterq.onsuccess = function(e) { done(); }
 
@@ -109,7 +109,7 @@ describe('W3C IDBFactory.deleteDatabase Tests', function () {
     it('IDBVersionChangeEvent fired in upgradeneeded, versionchange and deleteDatabase', function (done) {
         var db;
 
-        var openrq = indexedDB.open('db.sqlite', 3);
+        var openrq = indexedDB.open('db', 3);
 
         openrq.onupgradeneeded = function(e) {
             assert.equal(e.oldVersion, 0, "old version (upgradeneeded)");
@@ -139,7 +139,7 @@ describe('W3C IDBFactory.deleteDatabase Tests', function () {
         openrq.onblocked = function () { throw new Error("open.blocked"); };
 
         function deleteDB (e) {
-            var deleterq = indexedDB.deleteDatabase('db.sqlite');
+            var deleterq = indexedDB.deleteDatabase('db');
 
             deleterq.onsuccess = function(e) {
                 assert.equal(e.oldVersion, 3, "old version (delete.success)");