+++ tests-polyfill/w3c/TransactionBehavior.js
@@ -7,7 +7,7 @@ describe('W3C Transaction Behavior Tests', function () {
         var db, db_got_versionchange, db2,
             events = [];
 
-        var openrq = indexedDB.open('db.sqlite', 3);
+        var openrq = indexedDB.open('db', 3);
 
         // 1
         openrq.onupgradeneeded = function(e) {
@@ -46,7 +46,7 @@ describe('W3C Transaction Behavior Tests', function () {
             assert(db instanceof FDBDatabase);
             assert.deepEqual(db.objectStoreNames, [ "store" ]);
 
-            var openrq2 = indexedDB.open('db.sqlite', 4);
+            var openrq2 = indexedDB.open('db', 4);
 
             // 4
             openrq2.onupgradeneeded = function(e) {
@@ -75,7 +75,7 @@ describe('W3C Transaction Behavior Tests', function () {
                     ]);
 
                 if (db2) db2.close();
-                indexedDB.deleteDatabase('db.sqlite');
+                indexedDB.deleteDatabase('db');
                 setTimeout(function() {
                     done();
                 }, 10);
@@ -92,7 +92,7 @@ describe('W3C Transaction Behavior Tests', function () {
         var db, db_got_versionchange, db2,
             events = [];
 
-        var openrq = indexedDB.open('db.sqlite', 3);
+        var openrq = indexedDB.open('db', 3);
 
         // 1
         openrq.onupgradeneeded = function(e) {
@@ -131,7 +131,7 @@ describe('W3C Transaction Behavior Tests', function () {
             assert(db instanceof FDBDatabase);
             assert.deepEqual(db.objectStoreNames, [ "store" ]);
 
-            var openrq2 = indexedDB.open('db.sqlite', 4);
+            var openrq2 = indexedDB.open('db', 4);
 
             // 4
             openrq2.onblocked = function(e) {
@@ -168,7 +168,7 @@ describe('W3C Transaction Behavior Tests', function () {
                     ]);
 
                 if (db2) db2.close();
-                indexedDB.deleteDatabase('db.sqlite');
+                indexedDB.deleteDatabase('db');
                 setTimeout(function() { done(); }, 10);
             };