+++ tests-qunit/sampleData.js
@@ -1,7 +1,7 @@
 /* eslint-env qunit */
 /* eslint-disable no-var */
 var DB = { // eslint-disable-line no-unused-vars
-    NAME: 'dbname.sqlite',
+    NAME: 'dbname',
     OBJECT_STORE_1: 'objectStore1',
     OBJECT_STORE_2: 'objectStore2',
     OBJECT_STORE_3: 'objectStore3',